+++ src/tablecloth/pipeline.clj
@@ -1,7 +1,7 @@
 (ns tablecloth.pipeline
   "Linear pipeline operations."
   (:refer-clojure :exclude [group-by drop concat rand-nth first last shuffle])
-  (:require [tablecloth.api]))
+  (:require [tablecloth.api :as api]))
 
 (defmacro build-pipelined-function
   [f m]
@@ -10,7 +10,10 @@
        ~@(for [arg args
                :let [narg (mapv #(if (map? %) 'options %) arg)
                      [a & r] (split-with (partial not= '&) narg)]]
-           (list narg `(fn [ds#] (apply ~f ds# ~@a ~(rest r))))))))
+           (list narg `(fn [ds#]
+                         (let [ctx# (if (api/dataset? ds#)
+                                      {:metamorph/data ds#} ds#)]
+                           (assoc ctx# :metamorph/data (apply ~f (ctx# :metamorph/data) ~@a ~(rest r))))))))))
 
 (def ^:private excludes '#{dataset write-csv! let-dataset without-grouping->})
 
@@ -32,36 +35,62 @@
 (declare process-param)
 
 (defn- process-map
-  [ctx params]
+  [config params]
   (into {} (map (fn [[k v]]
-                  [k (process-param ctx v)]) params)))
+                  [k (process-param config v)]) params)))
 
 (defn- process-seq
-  [ctx params]
-  (mapv (fn [p] (process-param ctx p)) params))
+  [config params]
+  (mapv (fn [p] (process-param config p)) params))
+
+(defn- resolve-keyword
+  "Interpret keyword as a symbol and try to resolve it."
+  [k]
+  (-> (if-let [n (namespace k)] ;; namespaced?
+        (let [sn (symbol n)
+              n (str (get (ns-aliases *ns*) sn sn))] ;; try to find namespace in aliases
+          (symbol n (name k))) ;; create proper symbol with fixed namespace
+        (symbol (name k))) ;; no namespace case
+      (resolve)))
+
+(defn- maybe-var-get
+  "If symbol can be resolved, return var, else return original keyword"
+  [k]
+  (if-let [rk (resolve-keyword k)]
+    (var-get rk)
+    k))
 
 (defn- process-param
-  [ctx p]
+  "Recursively process parameters and try to resolve symbols for namespaced keywords.
+  Special case for namespaced keyword is `ctx` namespace. It means that we should look up in `config` map."
+  [config p]
   (cond
-    (and (keyword? p)
-         (namespace p)
-         (not (#{"type" "!type"} (namespace p)))) (let [n (namespace p)]
-                                                    (if (= n "ctx")
-                                                      (ctx (keyword (name p)))
-                                                      (var-get (resolve (symbol p)))))
-    (map? p) (process-map ctx p)
-    (sequential? p) (process-seq ctx p)
+    (and (keyword? p) ;; 
+         (let [n (namespace p)]
+           (and n (or (= n "ctx")
+                      (let [sn (symbol n)]
+                        (find-ns (get (ns-aliases *ns*) sn sn))))))) (let [n (namespace p)]
+                                                                       (if (= n "ctx")
+                                                                         (config (keyword (name p)))
+                                                                         (maybe-var-get p)))
+    (map? p) (process-map config p)
+    (sequential? p) (process-seq config p)
     :else p))
 
 (defn ->pipeline
+  "Create pipeline from declarative description."
   ([ops] (->pipeline {} ops))
-  ([ctx ops]
-   (apply pipeline (for [[op & params] ops
-                         :let [nparams (process-param ctx params)
-                               v (cond
-                                   (and (keyword? op)
-                                        (not (namespace op))) (var-get (resolve (symbol "tablecloth.pipeline" (name op))))
-                                   (keyword? op) (var-get (resolve (symbol op)))
-                                   (symbol? op) (resolve op)
-                                   :else op)]]
-                     (apply v nparams)))))
+  ([config ops]
+   (apply pipeline (for [line ops]
+                     (cond
+                       ;; if it's a sequence, resolve function, process parameters and call it.
+                       (sequential? line) (let [[op & params] line
+                                                nparams (process-param config params)
+                                                f (cond
+                                                    (keyword? op) (maybe-var-get op)
+                                                    (symbol? op) (var-get (resolve op))
+                                                    :else op)]
+                                            (apply f nparams))
+                       (keyword? line) (maybe-var-get line)
+                       :else line))))) ;; leave untouched otherwise
+